requirements.txt
Django>=2.2,<3.0
psycopg2==2.8.2
Pillow==6.0.0
django-widget-tweaks==1.4.3
django-debug-toolbar==1.11
sorl-thumbnail==12.5.0
graphene==2.1.5
graphene-django==2.2.0
django-filter==2.1.0
django-graphql-jwt==0.2.1